Use a single component to export
