--- src/pages/settings/wallet-settings/wallet-settings-advanced/wallet-backup-file/wallet-backup-file.scss
@@ -1,2 +1,0 @@
-page-wallet-backup-file {
-}
