+++ src/pages/settings/wallet-settings/wallet-settings.html
@@ -18,7 +18,8 @@
       </ion-item>
       <ion-item-divider>{{'Security' | translate}}</ion-item-divider>
       <button ion-item (click)="openBackupSettings()" [ngClass]="{warn: needsBackup}">
-        <span translate>Backup</span>
+        <span translate *ngIf="derivationStrategy != 'BIP45'">Backup</span>
+        <span translate *ngIf="derivationStrategy == 'BIP45'">Backup File</span>
         <ion-note *ngIf="needsBackup" class="wallet-warning" item-end>
           {{'Needs backup' | translate}}
         </ion-note>
@@ -50,7 +51,7 @@
       <button ion-item (click)="openWalletAddresses()">
         <span translate>Addresses</span>
       </button>
-      <button ion-item (click)="openExportWallet()">
+      <button ion-item *ngIf="derivationStrategy != 'BIP45'" (click)="openExportWallet()">
         <span translate>Export Wallet Metadata</span>
       </button>
       <button ion-item (click)="openWalletServiceUrl()">
