+++ src/pages/wallet-details/wallet-details.scss
@@ -385,7 +385,6 @@ page-bitpay-card-intro,
 page-bitpay-card,
 page-wallet-addresses,
 page-all-addresses,
-page-wallet-backup-file,
 page-wallet-color,
 page-wallet-delete,
 page-wallet-export,
